2reportsleft

+++ database/migrations/2023_10_06_110222_create_cash_received_vouchers_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('customer_id');
             $table->foreign('customer_id')->references('id')->on('customers');
             $table->string('desc');
-            $table->float('total_amount');
+            $table->bigInteger('total_amount');
             $table->timestamps();
         });
     }
